마무리

+++ src/main/java/com/banchango/warehouses/service/WarehousesService.java
@@ -19,15 +19,13 @@ import com.banchango.domain.warehouses.Warehouses;
 import com.banchango.domain.warehouses.WarehousesRepository;
 import com.banchango.domain.warehouseusagecautions.WarehouseUsageCautions;
 import com.banchango.tools.EmailContent;
+import com.banchango.users.exception.ForbiddenUserIdException;
 import com.banchango.users.exception.UserIdNotFoundException;
 import com.banchango.warehouses.dto.WarehouseDetailResponseDto;
 import com.banchango.warehouses.dto.WarehouseInsertRequestDto;
 import com.banchango.warehouses.dto.WarehouseSearchDto;
 import com.banchango.warehouses.dto.WarehouseUpdateRequestDto;
-import com.banchango.warehouses.exception.WarehouseIdNotFoundException;
-import com.banchango.warehouses.exception.WarehouseInvalidAccessException;
-import com.banchango.warehouses.exception.WarehouseNotFoundException;
-import com.banchango.warehouses.exception.WarehouseSearchException;
+import com.banchango.warehouses.exception.*;
 import lombok.RequiredArgsConstructor;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.data.domain.PageRequest;
@@ -75,6 +73,7 @@ public class WarehousesService {
                 .minReleasePerMonth(warehouseInsertRequestDto.getMinReleasePerMonth())
                 .latitude(warehouseInsertRequestDto.getLatitude())
                 .longitude(warehouseInsertRequestDto.getLongitude())
+                .status(WarehouseStatus.IN_PROGRESS)
                 .build();
 
         final Warehouses savedWarehouse = warehousesRepository.save(warehouse);
@@ -164,8 +163,16 @@ public class WarehousesService {
         return new WarehouseDetailResponseDto(warehouse, noImageUrl);
     }
 
-    public WarehouseDetailResponseDto updateWarehouse(Integer warehouseId, WarehouseUpdateRequestDto requestDto) {
+    @Transactional
+    public WarehouseDetailResponseDto updateWarehouse(String accessToken, Integer warehouseId, WarehouseUpdateRequestDto requestDto) {
+        int userId = JwtTokenUtil.extractUserId(accessToken);
+
         Warehouses warehouse = warehousesRepository.findById(warehouseId).orElseThrow(WarehouseIdNotFoundException::new);
+        if(warehouse.getStatus().equals(WarehouseStatus.DELETED)) throw new WarehouseNotFoundException();
+        if(!warehouse.getStatus().equals(WarehouseStatus.VIEWABLE)) throw new WarehouseIsNotViewableException();
+
+        if(!warehouse.getUserId().equals(userId)) throw new ForbiddenUserIdException();
+
         if(!mainItemTypesRepository.findByWarehouseId(warehouseId).equals(requestDto.getMainItemTypes())) {
             mainItemTypesRepository.deleteByWarehouseId(warehouseId);
         }
